Move executors into the same directory
