+++ src/coprocessor/dag/batch/executors/index_scan_executor.rs
@@ -20,18 +20,18 @@ use tipb::schema::ColumnInfo;
 
 use crate::storage::Store;
 
-use super::super::interface::*;
 use crate::coprocessor::codec::batch::{LazyBatchColumn, LazyBatchColumnVec};
+use crate::coprocessor::dag::batch::interface::*;
 use crate::coprocessor::dag::expr::{EvalConfig, EvalContext};
 use crate::coprocessor::dag::Scanner;
 use crate::coprocessor::{Error, Result};
 
 pub struct BatchIndexScanExecutor<C: ExecSummaryCollector, S: Store>(
-    super::scan_executor::ScanExecutor<
+    super::util::scan_executor::ScanExecutor<
         C,
         S,
         IndexScanExecutorImpl,
-        super::ranges_iter::PointRangeConditional,
+        super::util::ranges_iter::PointRangeConditional,
     >,
 );
 
@@ -59,7 +59,7 @@ impl<C: ExecSummaryCollector, S: Store> BatchIndexScanExecutor<C, S> {
         let mut columns_len_without_handle = 0;
         let mut decode_handle = false;
         for ci in &columns_info {
-            schema.push(super::scan_executor::field_type_from_column_info(&ci));
+            schema.push(super::util::scan_executor::field_type_from_column_info(&ci));
             if ci.get_pk_handle() {
                 decode_handle = true;
             } else {
@@ -73,13 +73,13 @@ impl<C: ExecSummaryCollector, S: Store> BatchIndexScanExecutor<C, S> {
             columns_len_without_handle,
             decode_handle,
         };
-        let wrapper = super::scan_executor::ScanExecutor::new(
+        let wrapper = super::util::scan_executor::ScanExecutor::new(
             summary_collector,
             imp,
             store,
             desc,
             key_ranges,
-            super::ranges_iter::PointRangeConditional::new(unique),
+            super::util::ranges_iter::PointRangeConditional::new(unique),
         )?;
         Ok(Self(wrapper))
     }
@@ -116,7 +116,7 @@ struct IndexScanExecutorImpl {
     decode_handle: bool,
 }
 
-impl super::scan_executor::ScanExecutorImpl for IndexScanExecutorImpl {
+impl super::util::scan_executor::ScanExecutorImpl for IndexScanExecutorImpl {
     #[inline]
     fn schema(&self) -> &[FieldType] {
         &self.schema
@@ -245,7 +245,7 @@ mod tests {
 
     use crate::coprocessor::codec::mysql::Tz;
     use crate::coprocessor::codec::{datum, table, Datum};
-    use crate::coprocessor::dag::batch_executor::statistics::*;
+    use crate::coprocessor::dag::batch::statistics::*;
     use crate::coprocessor::dag::expr::EvalConfig;
     use crate::coprocessor::util::convert_to_prefix_next;
     use crate::storage::{FixtureStore, Key};
