+++ src/coprocessor/dag/batch/executors/util/mod.rs
@@ -11,10 +11,5 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-mod index_scan_executor;
-mod ranges_iter;
-mod scan_executor;
-mod table_scan_executor;
-
-pub use self::index_scan_executor::BatchIndexScanExecutor;
-pub use self::table_scan_executor::BatchTableScanExecutor;
+pub mod ranges_iter;
+pub mod scan_executor;
