+++ src/coprocessor/dag/builder.rs
@@ -18,13 +18,13 @@ use tipb::executor::{self, ExecType};
 
 use crate::storage::Store;
 
-use super::batch_executor::executors::*;
-use super::batch_executor::interface::*;
+use super::batch::executors::*;
+use super::batch::interface::*;
 use super::executor::{
     Executor, HashAggExecutor, LimitExecutor, ScanExecutor, SelectionExecutor, StreamAggExecutor,
     TopNExecutor,
 };
-use crate::coprocessor::dag::batch_executor::statistics::ExecSummaryCollectorDisabled;
+use crate::coprocessor::dag::batch::statistics::ExecSummaryCollectorDisabled;
 use crate::coprocessor::dag::expr::EvalConfig;
 use crate::coprocessor::metrics::*;
 use crate::coprocessor::*;
