+++ src/coprocessor/dag/mod.rs
@@ -33,7 +33,7 @@
 //!
 //!   Obviously, this kind of executor must not be the first executor in the pipeline.
 
-pub mod batch_executor;
+pub mod batch;
 pub mod batch_handler;
 mod builder;
 pub mod executor;
